pyproject.toml
[project]
name = "rumi-leaf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "networkx>=3.5",
    "numpy>=2.3.4",
    "numpy-stl>=3.2.0",
    "pillow>=12.0.0",
    "pyfqmr>=0.5.0",
    "scikit-image>=0.25.2",
    "scipy>=1.16.3",
    "trimesh>=4.9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
